postgres_to_el_genres/requirements.txt
elasticsearch==8.11.0
psycopg2-binary==2.9.3
requests==2.32.5
pydantic ==2.11.9
datetime
pandas==2.3.3
redis
pydantic_settings==2.11.0
